re-enabled cherrypy.error log (fixes #120)
Our own logging config somehow silenced cherrypy.error (but not cherrypy.access for some reason). Not sure why it happens; but adding an explicit definition for cherrypy.error to our own configuration fixes it.

I used the opportunity to set the cherrypy loglevel to WARN, in order to keep the message volume down.

http://github.com/devsnd/cherrymusic/issues/issue/120