Configuration `in` test conforms to iteration (#94)
changed Configuration.__contains__ to no longer interpret its argument
as absolute key, but as relative to self. this conforms to the way
keys of children are listed during iteration.

__contains__ no longer accepts Properties as arguments, because its not
only unnecessary, but also has ambiguous semantics: does it just check
for the name or do a full equality check? bad dog! to the shelter with
you!