httphandler.api_listdir returns valid json on error (fixes #96)
the error message was not a valid json string. the resulting parse error
stopped script execution in the browser before the complete-handler could
hide() the waiting panel.

now, the error message gets handed to the list parser, who displays a
message about an "invalid server response" in the we client. not so
helpful, but at least the client doesn't lock up anymore.