Added working support for cuesheets.
This change adds support for cuesheets by passing a track parameter
around which indicates which track in the cuesheet should be played. The
actual audio file is split with ffmpeg using the -ss and -t parameters,
which are filled by the decoder when the file extension is '.cue'.

cuesheet.py is added to parse cuesheets. This is taken from a separate
project so it's a bit messy and contains some stuff that isn't used.

Metadata is also fetched from the cuesheet per track.