Pass metainfo from listdir between mediabrowser and playlist.
This allows passing of metainfo for cuesheets without calling
getsonginfo for every track. This gets rid of the starttime argument to
getsonginfo, and allows us to pass the metainfo we already have when
listing cuesheet contents without having to parse the cuesheet for
every track. getsonginfo should now be called only once for all tracks.